Update anchor generation for multi-link list-items
This looks related to #6.

The problem I'm seeing is that terraform's documentation will sometimes
contain list items with an <a> tag within the description of a <li>. The
previous regex would find that there was a link within the <li> already,
and then refuse to generate the anchor at the beginning of the <li>,
which made it hard for me to paste docs links to people in IRC :(.

I also added some tests for this case.