Add mega-nav
This is the first pass at adding the mega-nav. It's a three-step
process, but it unifies all the code under one place should we want to
make updates in the future.

If we like this approach, I'll need to make a new container and bump all
the projects. Let me know what you think.

/cc @captainill @mitchellh