Update supported values for pretty_os
Adds entry to properly format "archlinux" to "Arch Linux". Also
updates the linux match to only match on the explicit string, not
any value containing linux. This will prevent cases where a name
may not be defined (like archlinux) and it being displayed simply
as "Linux".

Also updated the pretty version of "darwin" to "macOS" since this
has been the official name since the Sierra release (2016).