Discard processed HTML if content is only wrapped in a paragraph
If there is no markdown content to be updated within the HTML
block the result will be the content wrapped with <p> tags. When
the rendered result is just the content wrapped in <p> tags,
assume no markdown content was found and return original value.
This prevents double processing happening in cases where a
partial may be used.